// pocketsync-web/api-client/model/conflict-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PocketSync API
 * API documentation for PocketSync backend services
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ConflictDto } from './conflict-dto';

/**
 * 
 * @export
 * @interface ConflictResponseDto
 */
export interface ConflictResponseDto {
    /**
     * Conflict details
     * @type {Array<ConflictDto>}
     * @memberof ConflictResponseDto
     */
    'conflicts': Array<ConflictDto>;
    /**
     * Total count of conflicts matching the filter criteria
     * @type {number}
     * @memberof ConflictResponseDto
     */
    'total': number;
    /**
     * Success status of the operation
     * @type {boolean}
     * @memberof ConflictResponseDto
     */
    'success': boolean;
}